+++ tests/testthat/test_plot_DetPlot.R
@@ -75,7 +75,7 @@ test_that("plot_DetPlot", {
 
   ## test self call with multi core
   SW({
-  results <- expect_s4_class(plot_DetPlot(
+  expect_message(expect_s4_class(plot_DetPlot(
     object = list(x = object, y = object),
     method = "shift",
     signal.integral.min = 1,
@@ -86,7 +86,25 @@ test_that("plot_DetPlot", {
       fit.method = "LIN",
       trim_channels = TRUE
     ),
-    multicore = 1,
+    multicore = 2,
+    n.channels = 2,
+    verbose = TRUE,
+    plot = FALSE),
+    "RLum.Results"),
+    "Running multicore session using 2 cores")
+
+  expect_s4_class(plot_DetPlot(
+    object = list(x = object, y = object),
+    method = "shift",
+    signal.integral.min = 1,
+    signal.integral.max = 3,
+    background.integral.min = 900,
+    background.integral.max = 1000,
+    analyse_function.control = list(
+      fit.method = "LIN",
+      trim_channels = TRUE
+    ),
+    multicore = FALSE,
     n.channels = 2,
     verbose = TRUE,
     plot = FALSE),